feat: move rag relevant interface to Application.Contract
