+++ src/AISmart.Application.Contracts/Chunk/IChunker.cs
@@ -1,7 +1,7 @@
 using System.Collections.Generic;
 using System.Threading.Tasks;
 
-namespace AISmart.Provider;
+namespace AISmart.Chunk;
 
 public interface IChunker {
     public Task<List<string>> Chunk(string text, int maxChunkSize);
