+++ src/AISmart.Application.Contracts/Rag/Dto/AddTextInput.cs
@@ -1,4 +1,4 @@
-namespace AISmart.Rag.dto;
+namespace AISmart.Rag.Dto;
 
 public class AddTextInput
 {
