--- src/AISmart.Application.Contracts/Rag/IChunker.cs
@@ -1,8 +1,0 @@
-using System.Collections.Generic;
-using System.Threading.Tasks;
-
-namespace AISmart.Rag;
-
-public interface IChunker {
-    public Task<List<string>> Chunk(string text, int maxChunkSize);
-}
