--- src/AISmart.Application.Contracts/Rag/IEmbeddingProvider.cs
@@ -1,8 +1,0 @@
-using System.Threading.Tasks;
-
-namespace AISmart.Rag;
-
-public interface IEmbeddingProvider
-{
-    Task<float[]> GetEmbeddingAsync(string text);
-}
