+++ src/AISmart.Application.Contracts/Rag/IRagProvider.cs
@@ -1,11 +1,11 @@
-// using System.Collections.Generic;
-// using System.Threading.Tasks;
-//
-// namespace AISmart.Rag;
-//
-// public interface IRagProvider
-// {
-//     Task StoreTextAsync(string text);
-//     Task BatchStoreTextsAsync(IEnumerable<string> texts);
-//     Task<string> RetrieveAnswerAsync(string query);
-// }
+using System.Collections.Generic;
+using System.Threading.Tasks;
+
+namespace AISmart.Rag;
+
+public interface IRagProvider
+{
+    Task StoreTextAsync(string text);
+    Task BatchStoreTextsAsync(IEnumerable<string> texts);
+    Task<string> RetrieveAnswerAsync(string query);
+}
