+++ src/AISmart.HttpApi/Controllers/RagController.cs
@@ -1,11 +1,8 @@
 using System.Threading.Tasks;
-using AISmart.Options;
-using AISmart.Provider;
-using AISmart.Rag.dto;
+using AISmart.Rag;
+using AISmart.Rag.Dto;
 using Asp.Versioning;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.Extensions.Logging;
-using Microsoft.Extensions.Options;
 using Volo.Abp;
 
 namespace AISmart.Controllers;
