+++ src/AISmart.Rag/Provider/RagProvider.cs
@@ -2,7 +2,11 @@ using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Threading.Tasks;
+using AISmart.Chunk;
+using AISmart.Embedding;
 using AISmart.Options;
+using AISmart.Rag;
+using AISmart.VectorStorage;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 using Volo.Abp.DependencyInjection;
